+++ packages/cli/src/index.ts
@@ -6,21 +6,25 @@ import {
   readFile,
   writeToFile,
   PossibleSchemaInput,
+  safeJSONParse,
 } from '@gql2ts/util';
 import { ISchemaToInterfaceOptions, generateNamespace } from '@gql2ts/from-schema';
 import fromQuery from '@gql2ts/from-query';
 import { IFromQueryReturnValue } from '@gql2ts/types';
+
+// tslint:disable-next-line no-require-imports no-var-requires
 const { version } = require('../package.json');
 
 program
   .version(version)
-  .usage('[options] <schema.json> <query.gql>')
+  .usage('[options] <schema.json | schema.gql> <query.gql>')
   .option('-o --output-file [outputFile]', 'name for output file, will use stdout if not specified')
   .option('-n --namespace [namespace]', 'name for the namespace, defaults to "GQL"', 'GQL')
   .option('-i --ignored-types <ignoredTypes>', 'names of types to ignore (comma delimited)', v => v.split(','), [])
   .option('-l --legacy', 'Use TypeScript 1.x annotation', false)
   .option('-e --external-options [externalOptions]', 'ES Module with method overwrites')
   .option('--ignore-type-name-declaration', 'Whether to exclude __typename', false)
+  .option('--exclude-deprecated-fields', 'Whether to exclude deprecated fields', false)
   .parse(process.argv);
 
 interface ICLIOptions extends Partial<ISchemaToInterfaceOptions> {
@@ -31,7 +35,8 @@ const run: (schema: PossibleSchemaInput, options: Partial<ICLIOptions>) => void 
   let defaultOverrides: object = {};
   if (program.externalOptions) {
     // tslint:disable-next-line no-require-imports no-var-requires
-    defaultOverrides = require(program.externalOptions);
+    const externalFile: any = require(program.externalOptions);
+    defaultOverrides = externalFile.default || externalFile;
   }
 
   if (program.args[1]) {
@@ -58,17 +63,17 @@ const run: (schema: PossibleSchemaInput, options: Partial<ICLIOptions>) => void 
 
 const fileName: string | undefined = program.args[0];
 
-if (!process.stdin.isTTY) {
+if (fileName) {
+  const schema: string | object = readFile(fileName);
+  run(schema as PossibleSchemaInput, program as any);
+} else if (!process.stdin.isTTY) {
   let input: string = '';
   process.stdin.resume();
   process.stdin.setEncoding('utf8');
   process.stdin.on('data', (data) => {
     input += data;
   });
-  process.stdin.on('end', () => run(JSON.parse(input), program as any));
-} else if (fileName) {
-  const schema: string = readFile(fileName);
-  run(schema, program as any);
+  process.stdin.on('end', () => run(safeJSONParse(input) as PossibleSchemaInput, program as any));
 } else {
   console.error('No input specified. Please use stdin or a file name.');
   program.outputHelp();