+++ packages/from-schema/src/index.ts
@@ -42,7 +42,7 @@ const run: (schemaInput: GraphQLSchema, optionsInput: IInternalOptions) => strin
     addSemicolon,
   } = optionsInput.formats;
 
-  const TYPE_MAP: ITypeMap = { ...DEFAULT_TYPE_MAP };
+  const TYPE_MAP: ITypeMap = { ...DEFAULT_TYPE_MAP, ...(optionsInput.typeMap || {}) };
 
   function isScalar (type: any): type is GraphQLScalarType {
     return type instanceof GraphQLScalarType || !!(type as any)._scalarConfig;
@@ -84,11 +84,40 @@ const run: (schemaInput: GraphQLSchema, optionsInput: IInternalOptions) => strin
     line: number;
     column: number;
   }`;
+  type GenerateDescription = (description?: string, jsDoc?: IJSDocTag[]) => string;
+  const generateDescription: GenerateDescription = (description, jsDoc = []) => (description || jsDoc.length) ? `/**
+    ${[description, ...jsDoc.map(({ tag, value }) => `@${tag} ${value}`)].filter(x => !!x).join('\n')}
+  */` : '';
 
-  const wrapWithDescription: (declaration: string, description: string) => string = (declaration, description) => `  /*
-    description: ${description}
-  */
+  const wrapWithDescription: (declaration: string, description: string) => string = (declaration, description) =>
+  `  ${generateDescription(description)}
   ${declaration}`;
+
+  interface IJSDocTag {
+    tag: string;
+    value: string;
+  }
+
+  function isInputField (field: GraphQLField<any, any> | GraphQLInputField): field is GraphQLInputField {
+    return !!field.astNode && field.astNode.kind === 'InputValueDefinition';
+  }
+
+  const buildDocTags: (field: GraphQLField<any, any> | GraphQLInputField) => IJSDocTag[] = field => {
+    const tags: IJSDocTag[] = [];
+    if (!field.astNode) {
+      return tags;
+    } else if (isInputField(field)) {
+      if (field.defaultValue) {
+        tags.push({ tag: 'default', value: field.defaultValue });
+      }
+    } else {
+      if (field.isDeprecated) {
+        tags.push({ tag: 'deprecated', value: field.deprecationReason || '' });
+      }
+    }
+
+    return tags;
+  };
 
   const generateTypeDeclaration: (description: string, name: string, possibleTypes: string) => string =
     (description, name, possibleTypes) => wrapWithDescription(addSemicolon(typeBuilder(name, possibleTypes)) + '\n\n', description);
@@ -162,7 +191,7 @@ const run: (schemaInput: GraphQLSchema, optionsInput: IInternalOptions) => strin
 
   const filterField: (field: GraphQLField<any, any> | GraphQLInputField, ignoredTypes: Set<String>) => boolean = (field, ignoredTypes) => {
     let nestedType: GraphQLNamedType = findRootType(field.type);
-    return !ignoredTypes.has(nestedType.name);
+    return !ignoredTypes.has(nestedType.name) && (!optionsInput.excludeDeprecatedFields || !(field as GraphQLField<any, any>).isDeprecated);
   };
 
   type InterfaceMap = Map<GraphQLInterfaceType, GraphQLObjectType[]>;
@@ -208,7 +237,8 @@ const run: (schemaInput: GraphQLSchema, optionsInput: IInternalOptions) => strin
 
     let fields: string[] = f
       .filter(field => filterField(field, ignoredTypes))
-      .map(field => fieldToDefinition(field, isInput, supportsNullability))
+      .map(field => [generateDescription(field.description, buildDocTags(field)), fieldToDefinition(field, isInput, supportsNullability)])
+      .reduce((acc, val) => [...acc, ...val.filter(x => x)] , [])
       .filter(field => field);
 
     let interfaceDeclaration: string = generateInterfaceName(type.name);
@@ -297,6 +327,8 @@ export interface ISchemaToInterfaceOptions {
   namespace: string;
   outputFile?: string;
   externalOptions?: string;
+  typeMap?: ITypeMap;
+  excludeDeprecatedFields?: boolean;
 }
 
 export interface IInternalOptions extends Partial<ISchemaToInterfaceOptions> {